Right flight path
Rocket now takes of in the right direction (e.i. up instead of down), and z, v, and a go to zero if z is zero. I also cleared most of Forrest's comment, I though there weren't that useful anymore but we can bring them back if you want to.